+++ hemi/develop/svcs/diogin/service.go
@@ -3,17 +3,14 @@
 // All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be found in the LICENSE.md file.
 
-// This is a hello svc showing how to use Gorox RPC Framework to host a svc.
-
-package hello
+package diogin
 
 import (
 	. "github.com/hexinfra/gorox/hemi"
 )
 
 func init() {
-	// Register svc initializer.
-	RegisterSvcInit("hello", func(svc *Svc) error {
+	RegisterServiceInit("diogin", func(service *Service) error {
 		return nil
 	})
 }