+++ hemi/develop/test/apps/diogin/test.go
@@ -18,7 +18,7 @@ func Main() {
 	hemi.SetTmpsDir(system.ExeDir + "/tmps")
 	hemi.SetVarsDir(system.ExeDir + "/vars")
 
-	stage, err := hemi.FromText("stage{}")
+	stage, err := hemi.BootText("stage{}")
 	if err != nil {
 		fmt.Println(err.Error())
 		return