+++ hemi/procman/procman.go
@@ -160,7 +160,7 @@ func Main(args *Args) {
 		setDir(&common.VarsDir, "vars", hemi.SetVarsDir)
 
 		if action == "check" { // dry run
-			if _, err := hemi.FromFile(common.GetConfig()); err != nil {
+			if _, err := hemi.BootFile(common.GetConfig()); err != nil {
 				fmt.Fprintln(os.Stderr, err.Error())
 			} else {
 				fmt.Println("PASS")
@@ -170,7 +170,7 @@ func Main(args *Args) {
 
 		// Now serve.
 		if common.SingleMode { // run as single foreground process. for single mode
-			if stage, err := hemi.FromFile(common.GetConfig()); err == nil {
+			if stage, err := hemi.BootFile(common.GetConfig()); err == nil {
 				stage.Start(0)
 				select {} // waiting forever
 			} else {